backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
apscheduler
aiohttp
requests
pydantic
lxml
python-dotenv
yfinance
